angular-map/backend/requirements.txt
Flask
Flask-Cors
werkzeug
psycopg2-binary
numpy
scikit-learn
pandas
requests